Move experimental features into the main client
